Consolidate the 3 SharedDictionaryManager classes
Everywhere now uses `GitHub.UI.ThemeDictionaryManager`.
SharedDictionaryManager can be used as base for added functionality (e.g. ThemeDictionaryManager).
