--- src/GitHub.VisualStudio.UI/Helpers/SharedDictionaryManager.cs
@@ -1,8 +1,0 @@
-﻿using GitHub.UI.Helpers;
-
-namespace GitHub.VisualStudio.UI.Helpers
-{
-    public class SharedDictionaryManager : SharedDictionaryManagerBase
-    {
-    }
-}
