+++ src/GitHub.VisualStudio.UI/Styles/LinkDropDown.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:cache="clr-namespace:GitHub.VisualStudio.UI.Helpers"
+                    xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                     xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI">
 
     <ResourceDictionary.MergedDictionaries>
