+++ src/GitHub.VisualStudio/GitHub.VisualStudio.csproj
@@ -297,7 +297,6 @@
     <Compile Include="Menus\BlameLink.cs" />
     <Compile Include="Menus\MenuBase.cs" />
     <Compile Include="Menus\CreateGist.cs" />
-    <Compile Include="Helpers\SharedDictionaryManager.cs" />
     <Compile Include="Helpers\ActiveDocumentSnapshot.cs" />
     <Compile Include="Menus\OpenLink.cs" />
     <Compile Include="Menus\LinkMenuBase.cs" />
