--- src/GitHub.VisualStudio/Helpers/SharedDictionaryManager.cs
@@ -1,8 +1,0 @@
-﻿using GitHub.UI.Helpers;
-
-namespace GitHub.VisualStudio.Helpers
-{
-    public class SharedDictionaryManager : SharedDictionaryManagerBase
-    {
-    }
-}
