+++ src/GitHub.VisualStudio/Resources/icons/mark_github_toolbar.xaml
@@ -1,7 +1,7 @@
 ﻿<Viewbox xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
          Width="1024" Height="1024">
     <Border BorderBrush="Transparent" BorderThickness="1">
-        <ui:OcticonImage xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+        <ui:OcticonImage xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                      xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI"
                      Foreground="#00458D"
                      Icon="mark_github">
