+++ src/GitHub.VisualStudio/UI/Views/Controls/LoginControl.xaml
@@ -6,7 +6,7 @@
              xmlns:ui="clr-namespace:GitHub.UI;assembly=GitHub.UI"
              xmlns:uirx="clr-namespace:GitHub.UI;assembly=GitHub.UI.Reactive"
              xmlns:helpers="clr-namespace:GitHub.Helpers;assembly=GitHub.UI"
-             xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+             xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
              xmlns:prop="clr-namespace:GitHub.VisualStudio.UI;assembly=GitHub.VisualStudio.UI"
              xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views.Controls"
              xmlns:automation="clr-namespace:GitHub.UI.TestAutomation;assembly=GitHub.UI"
