+++ src/GitHub.VisualStudio/UI/Views/Controls/PullRequestListItem.xaml
@@ -1,6 +1,6 @@
 ﻿<ResourceDictionary xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                    xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+                    xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                     xmlns:i18n="clr-namespace:GitHub.VisualStudio.UI;assembly=GitHub.VisualStudio.UI"
                     xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views"
                     xmlns:models="clr-namespace:GitHub.Models;assembly=GitHub.Exports"
