+++ src/GitHub.VisualStudio/UI/Views/Controls/RepositoryCloneControl.xaml
@@ -2,7 +2,7 @@
                             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
                             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-                            xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+                            xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                             xmlns:helpers="clr-namespace:GitHub.Helpers;assembly=GitHub.UI"
                             xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views.Controls"
