+++ src/GitHub.VisualStudio/UI/Views/Controls/RepositoryPublishControl.xaml
@@ -1,9 +1,9 @@
 ﻿<local:GenericRepositoryPublishControl x:Class="GitHub.VisualStudio.UI.Views.Controls.RepositoryPublishControl"
                                        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                                        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                                       xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+                                       xmlns:helpers="clr-namespace:GitHub.VisualStudio.Helpers"
                                        xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-                                       xmlns:helpers="clr-namespace:GitHub.Helpers;assembly=GitHub.UI"
+                                       xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                                        xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views.Controls"
                                        xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                                        xmlns:prop="clr-namespace:GitHub.VisualStudio.UI;assembly=GitHub.VisualStudio.UI"
@@ -110,7 +110,7 @@
                Margin="0"
                Background="{DynamicResource GitHubVsSearchBoxBackground}"
                Foreground="{DynamicResource GitHubVsWindowText}"
-               MaxLength="{x:Static cache:Constants.MaxRepositoryNameLength}"
+               MaxLength="{x:Static helpers:Constants.MaxRepositoryNameLength}"
                Text="{Binding RepositoryName}"
                PromptText="{x:Static prop:Resources.RepoNameText}" 
                AutomationProperties.AutomationId="{x:Static automation:AutomationIDs.TeamExplorerPublishRepositoryNameTextBox}"
