+++ src/GitHub.VisualStudio/UI/Views/Controls/TwoFactorControl.xaml
@@ -8,7 +8,7 @@
     xmlns:local="clr-namespace:GitHub.VisualStudio.UI.Views.Controls"
     xmlns:helpers="clr-namespace:GitHub.Helpers;assembly=GitHub.UI"
     xmlns:sampleData="clr-namespace:GitHub.SampleData;assembly=GitHub.App"
-    xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+    xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
     xmlns:prop="clr-namespace:GitHub.VisualStudio.UI;assembly=GitHub.VisualStudio.UI"
     xmlns:automation="clr-namespace:GitHub.UI.TestAutomation;assembly=GitHub.UI"
     mc:Ignorable="d"
