+++ src/GitHub.VisualStudio/UI/Views/GitHubPaneView.xaml
@@ -1,7 +1,7 @@
 ﻿<view:GenericGitHubPaneView x:Class="GitHub.VisualStudio.UI.Views.GitHubPaneView"
                             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
                             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-                            xmlns:cache="clr-namespace:GitHub.VisualStudio.Helpers"
+                            xmlns:cache="clr-namespace:GitHub.UI.Helpers;assembly=GitHub.UI"
                             xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
                             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
                             xmlns:opts="http://schemas.microsoft.com/winfx/2006/xaml/presentation/options"
